+++ src/main/java/model/state/finished/Bust.java
@@ -8,10 +8,10 @@ public class Bust extends Finished {
     }
 
     @Override
-    public double profit(final int bettingMoney, final Cards dealerCards) {
+    public int profit(final int bettingMoney, final Cards dealerCards) {
         if(dealerCards.isHigherThan21()) {
-            return 0.0;
+            return 0;
         }
-        return bettingMoney * -1.0;
+        return bettingMoney * -1;
     }
 }
