+++ src/main/java/model/state/running/Running.java
@@ -14,7 +14,7 @@ public abstract class Running extends Created {
     }
 
     @Override
-    public double profit(final int money, final Cards dealerCards) {
+    public int profit(final int money, final Cards dealerCards) {
         throw new IllegalArgumentException("아직 Hit 상태라 수익을 알 수 없습니다.");
     }
 }
